refactor(api): extract bearer-auth into Collavre::Api::Authenticatable concern
