+++ engines/collavre_completion_api/app/controllers/collavre_completion_api/api/v1/base_controller.rb
@@ -4,6 +4,8 @@ module CollavreCompletionApi
   module Api
     module V1
       class BaseController < ActionController::API
+        include Collavre::Api::Authenticatable
+
         before_action :authenticate!
 
         private
@@ -17,34 +19,12 @@ module CollavreCompletionApi
             return
           end
 
-          unless authenticate_oauth(token)
+          unless find_user_by_bearer_token(token)
             render json: { error: { message: "Invalid authentication token", type: "invalid_request_error",
                                     code: "invalid_token" } },
                    status: :unauthorized
             nil
           end
-        end
-
-        def authenticate_oauth(token)
-          access_token = Doorkeeper::AccessToken.by_token(token)
-          return false unless access_token&.accessible?
-
-          user = Collavre::User.find_by(id: access_token.resource_owner_id)
-          return false unless user
-
-          Collavre::Current.user = user
-          true
-        end
-
-        def extract_bearer_token
-          auth_header = request.headers["Authorization"]
-          return nil unless auth_header&.start_with?("Bearer ")
-
-          auth_header.sub("Bearer ", "")
-        end
-
-        def current_user
-          Collavre::Current.user
         end
 
         # Model ID format: email username (before @)
